# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "description": "The error code vocabulary (IMPLEMENTATION_PLAN section 13).\n\nTwo codes need a note:\n\n* ``RATE_LIMITED`` is part of the published contract but is never raised in\n  Phase 1 — no rate limiting is implemented yet.\n* ``VALIDATION_ERROR`` is an addition to section 13's list (which the plan\n  calls \"主なエラーコード\", i.e. non-exhaustive). Request validation failures\n  that are not about a path or a title — a non-integer ``limit``, an unknown\n  body field — need a code of their own rather than being forced into\n  ``INVALID_PATH``.",
        "enum": [
          "UNAUTHORIZED",
          "INVALID_PATH",
          "PATH_OUTSIDE_VAULT",
          "NOTE_NOT_FOUND",
          "INVALID_FILE_TYPE",
          "FILE_TOO_LARGE",
          "INVALID_TITLE",
          "NOTE_ALREADY_EXISTS",
          "VALIDATION_ERROR",
          "RATE_LIMITED",
          "INVALID_CURSOR",
          "NOTE_MODIFIED",
          "INBOX_LOCK_TIMEOUT",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The single error shape used by every failing response (section 13).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque pagination token. Pass it back as `cursor` with the same query/folder/tags to fetch the next page. Null when there is no further page.",
            "title": "Next Cursor"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "title": "Results",
            "type": "array"
          },
          "skipped_count": {
            "default": 0,
            "description": "Notes that matched the requested scope but could not be read while scanning the Vault for this search, and were excluded from `results` as a result — never the difference between `results` and the total number of matches, which is what `next_cursor` accounts for. 0 on an ordinary search.",
            "minimum": 0.0,
            "title": "Skipped Count",
            "type": "integer"
          }
        },
        "required": [
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchResultItem": {
        "properties": {
          "excerpt": {
            "description": "Short plain-text snippet around the match.",
            "title": "Excerpt",
            "type": "string"
          },
          "id": {
            "description": "Vault-relative path; pass it to readNote as `path`.",
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "description": "File mtime in the configured timezone.",
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "path": {
            "description": "Vault-relative path of the note.",
            "title": "Path",
            "type": "string"
          },
          "tags": {
            "description": "Frontmatter tags, in file order.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "description": "Frontmatter `title`, else the file name without .md.",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "title",
          "excerpt",
          "tags",
          "modified_at"
        ],
        "title": "SearchResultItem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Send `Authorization: Bearer <API_TOKEN>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-mostly gateway over an Obsidian vault: full-vault search, note reads, vault tree/summary browsing, inbox duplicate-candidate detection, and note creation and append restricted to 00_Inbox/ChatGPT. MCP (mounted at /mcp) is the primary interface; this REST API is kept for health checks, curl-based diagnostics, and regression tests. Every failing response uses the single `{\"error\": {\"code\": ..., \"message\": ...}}` envelope described on each operation below — never FastAPI's default `{\"detail\": ...}` shape.",
    "title": "Obsidian Vault Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "operationId": "searchNotes",
        "parameters": [
          {
            "description": "Free-text search term.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search term.",
              "title": "Q"
            }
          },
          {
            "description": "Vault-relative folder to restrict the search to.",
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Vault-relative folder to restrict the search to.",
              "title": "Folder"
            }
          },
          {
            "description": "Comma-separated frontmatter tags; all must match.",
            "in": "query",
            "name": "tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated frontmatter tags; all must match.",
              "title": "Tags"
            }
          },
          {
            "description": "Maximum number of results.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Maximum number of results.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque pagination token from a previous response.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination token from a previous response.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INVALID_PATH, INVALID_CURSOR, VALIDATION_ERROR."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: UNAUTHORIZED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INTERNAL_ERROR."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every failing response — including any status code not listed individually above — uses this same envelope; only `error.code` and `error.message` vary."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Search the vault by file name, title, tags, headings and body",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}